@@ -28,7 +28,7 @@
     "fix:prettier": "prettier \"{,!(dist|temp|bundle)/**/}*.{ts,tsx}\" --write"
   },
   "dependencies": {
-    "@material-ui/core": "^4.0.1",
+    "@material-ui/core": "^4.3.1",
     "@material-ui/icons": "^4.0.1",
     "@sensenet/authentication-google": "^2.0.10",
     "@sensenet/authentication-jwt": "^1.0.14",
@@ -43,11 +43,11 @@
     "@sensenet/redux": "^5.1.11",
     "@sensenet/repository-events": "^1.4.3",
     "@sensenet/search-react": "^1.2.7",
-    "autoprefixer": "^9.5.1",
-    "css-loader": "^3.0.0",
+    "autoprefixer": "^9.6.1",
+    "css-loader": "^3.1.0",
     "dotenv": "^8.0.0",
-    "file-loader": "4.0.0",
-    "fs-extra": "^8.0.1",
+    "file-loader": "4.1.0",
+    "fs-extra": "^8.1.0",
     "history": "^4.7.2",
     "html-webpack-plugin": "^3.2.0",
     "lodash.groupby": "^4.6.0",
@@ -57,11 +57,11 @@
     "postcss-loader": "^3.0.0",
     "promise": "^8.0.3",
     "react": "^16.8.4",
-    "react-async-script": "^1.0.1",
-    "react-cookie-consent": "^2.3.1",
+    "react-async-script": "^1.1.1",
+    "react-cookie-consent": "^2.3.2",
     "react-custom-scrollbars": "^4.2.1",
     "react-dom": "^16.8.4",
-    "react-google-recaptcha": "^1.0.5",
+    "react-google-recaptcha": "^1.1.0",
     "react-loadable": "^5.5.0",
     "react-markdown": "^4.0.8",
     "react-moment": "^0.9.2",
@@ -69,47 +69,47 @@
     "react-responsive": "^7.0.0",
     "react-router": "^5.0.0",
     "react-router-dom": "^5.0.0",
-    "redux": "^4.0.1",
+    "redux": "^4.0.4",
     "redux-di-middleware": "^4.0.1",
     "reflect-metadata": "^0.1.12",
     "semaphore-async-await": "^1.5.1",
     "ts-keycode-enum": "^1.0.6",
-    "typeface-roboto": "^0.0.54",
+    "typeface-roboto": "^0.0.75",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
     "@cypress/webpack-preprocessor": "^4.0.3",
-    "@types/chance": "^1.0.4",
-    "@types/jest": "^24.0.9",
+    "@types/chance": "^1.0.5",
+    "@types/jest": "^24.0.16",
     "@types/lodash.groupby": "^4.6.6",
-    "@types/node": "^12.0.3",
-    "@types/react": "^16.8.19",
+    "@types/node": "^12.6.8",
+    "@types/react": "^16.8.23",
     "@types/react-custom-scrollbars": "^4.0.5",
-    "@types/react-dom": "^16.8.4",
-    "@types/react-google-recaptcha": "^1.0.0",
+    "@types/react-dom": "^16.8.5",
+    "@types/react-google-recaptcha": "^1.1.0",
     "@types/react-loadable": "^5.4.2",
-    "@types/react-redux": "^7.0.9",
+    "@types/react-redux": "^7.1.1",
     "@types/react-responsive": "^3.0.2",
     "@types/react-router-dom": "^4.3.3",
     "@types/redux-mock-store": "^1.0.1",
-    "@types/uuid": "^3.4.4",
+    "@types/uuid": "^3.4.5",
     "awesome-typescript-loader": "^5.2.1",
     "chance": "^1.0.18",
-    "copy-webpack-plugin": "^5.0.3",
+    "copy-webpack-plugin": "^5.0.4",
     "cross-env": "^5.2.0",
-    "cypress": "^3.3.1",
+    "cypress": "^3.4.1",
     "path-to-regexp": "3.0.0",
-    "raw-loader": "^3.0.0",
+    "raw-loader": "^3.1.0",
     "redux-mock-store": "^1.5.3",
     "source-map-loader": "^0.2.4",
     "style-loader": "^0.23.1",
-    "svg-url-loader": "^2.3.2",
+    "svg-url-loader": "^3.0.0",
     "uglify-es": "^3.3.9",
     "uglifyjs-webpack-plugin": "^2.1.3",
-    "url-loader": "^2.0.1",
-    "webpack": "^4.32.2",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.3.2",
+    "url-loader": "^2.1.0",
+    "webpack": "^4.38.0",
+    "webpack-bundle-analyzer": "^3.4.1",
+    "webpack-cli": "^3.3.6",
     "webpack-dev-server": "^3.4.1"
   }
 }
