@@ -41,10 +41,10 @@
     "react": "^16.8.4",
     "react-dom": "^16.8.4",
     "react-redux": "^7.0.3",
-    "redux": "^4.0.1"
+    "redux": "^4.0.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.5",
+    "@babel/core": "^7.5.5",
     "@storybook/addon-a11y": "^5.0.11",
     "@storybook/addon-actions": "^5.0.11",
     "@storybook/addon-events": "^5.0.11",
@@ -57,21 +57,21 @@
     "@storybook/addon-viewport": "^5.0.11",
     "@storybook/addons": "^5.0.11",
     "@storybook/react": "^5.0.11",
-    "@types/node": "^12.0.3",
-    "@types/react": "^16.8.19",
-    "@types/react-redux": "^7.0.9",
+    "@types/node": "^12.6.8",
+    "@types/react": "^16.8.23",
+    "@types/react-redux": "^7.1.1",
     "@types/storybook__addon-a11y": "^5.0.0",
     "@types/storybook__addon-actions": "^3.4.2",
     "@types/storybook__addon-info": "^4.1.0",
     "@types/storybook__addon-options": "^4.0.1",
     "@types/storybook__addon-viewport": "^4.1.0",
     "@types/storybook__react": "^4.0.1",
-    "@types/uuid": "^3.4.4",
+    "@types/uuid": "^3.4.5",
     "babel-loader": "^8.0.6",
     "babel-preset-react-app": "^9.0.0",
     "react-docgen-typescript-loader": "^3.0.1",
-    "ts-config-webpack-plugin": "^1.3.1",
+    "ts-config-webpack-plugin": "^1.4.0",
     "uuid": "^3.3.2",
-    "webpack": "^4.32.2"
+    "webpack": "^4.38.0"
   }
 }
