@@ -1,45 +1,44 @@
 {
   "private": true,
   "version": "2019.7.0",
-  "workspaces": {
-    "packages": [
-      "packages/*",
-      "examples/*",
-      "apps/*"
-    ]
-  },
+  "workspaces": [
+    "packages/*",
+    "examples/*",
+    "apps/*"
+  ],
   "config": {
     "commitizen": {
-      "path": "sensenet-kfi-cz-conventional-changelog"
+      "path": "cz-conventional-changelog"
     }
   },
   "devDependencies": {
-    "@types/enzyme": "^3.9.3",
-    "@types/jest": "^24.0.13",
-    "@typescript-eslint/eslint-plugin": "^1.10.2",
-    "@typescript-eslint/parser": "^1.10.2",
+    "@types/enzyme": "^3.10.3",
+    "@types/jest": "^24.0.16",
+    "@typescript-eslint/eslint-plugin": "^1.13.0",
+    "@typescript-eslint/parser": "^1.13.0",
+    "cz-conventional-changelog": "^3.0.2",
     "enzyme": "^3.9.0",
     "enzyme-adapter-react-16": "^1.13.2",
     "enzyme-to-json": "^3.3.5",
-    "eslint": "^6.0.1",
+    "eslint": "^6.1.0",
     "eslint-config-prettier": "^6.0.0",
-    "eslint-plugin-cypress": "^2.2.1",
-    "eslint-plugin-import": "^2.17.3",
-    "eslint-plugin-jsdoc": "^8.0.0",
+    "eslint-plugin-cypress": "^2.6.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsdoc": "^15.8.0",
     "eslint-plugin-prettier": "^3.1.0",
-    "eslint-plugin-react": "^7.13.0",
+    "eslint-plugin-react": "^7.14.3",
     "eslint-plugin-react-hooks": "^1.6.0",
-    "husky": "^2.3.0",
+    "husky": "^3.0.2",
     "jest": "^24.8.0",
-    "lerna": "^3.14.1",
-    "lint-staged": "^8.1.7",
+    "lerna": "^3.16.4",
+    "lint-staged": "^9.2.1",
     "prettier": "^1.17.1",
     "rimraf": "^2.6.3",
     "start-server-and-test": "^1.9.1",
     "ts-jest": "^24.0.2",
     "tslib": "^1.9.3",
-    "typescript": "^3.5.1",
-    "wait-on": "^3.2.0"
+    "typescript": "^3.5.3",
+    "wait-on": "^3.3.0"
   },
   "scripts": {
     "cy:run:dms": "cypress run --project ./examples/sn-dms-demo",
@@ -64,19 +63,14 @@
     }
   },
   "lint-staged": {
-    "linters": {
-      "*.{ts,tsx}": [
-        "eslint --ext .tsx,.ts --cache --fix",
-        "prettier --write",
-        "git add"
-      ],
-      "*.{js,jsx}": [
-        "prettier --write",
-        "git add"
-      ]
-    },
-    "ignore": [
-      "*.d.ts"
+    "*.{ts,tsx}": [
+      "eslint --ext .tsx,.ts --cache --fix",
+      "prettier --write",
+      "git add"
+    ],
+    "*.{js,jsx}": [
+      "prettier --write",
+      "git add"
     ]
   },
   "jest": {
