@@ -1,7 +1,7 @@
 import ClickAwayListener from '@material-ui/core/ClickAwayListener'
 import TextField from '@material-ui/core/TextField'
 import Typography from '@material-ui/core/Typography'
-import React from 'react'
+import React, { useState } from 'react'
 import { connect } from 'react-redux'
 import { debounce } from '@sensenet/client-utils'
 import { RootReducerType, setActivePages } from '../../store'
@@ -28,122 +28,65 @@ export const mapDispatchToProps = {
 }
 
 /**
- * Defines the own props for the PagerState component
- */
-export interface PagerState {
-  currentPage: number
-  lastPage: number
-  focused: boolean
-}
-
-/**
  * Document widget component for paging
  */
-export class DocumentTitlePagerComponent extends React.Component<
-  ReturnType<typeof mapStateToProps> & typeof mapDispatchToProps,
-  PagerState
-> {
-  /** the component state */
-  public state: PagerState = {
-    currentPage: this.props.activePages[0] || 1,
-    lastPage: this.props.pageCount,
-    focused: false,
-  }
+export const DocumentTitlePagerComponent: React.FC<
+  ReturnType<typeof mapStateToProps> & typeof mapDispatchToProps
+> = props => {
+  const [isFocused, setIsFocused] = useState(false)
 
-  private setPage = debounce(() => {
-    this.props.setActivePages([this.state.currentPage])
-  }, 200).bind(this)
+  const setPage = debounce((index: number) => {
+    props.setActivePages([index])
+  }, 200)
 
-  /** triggered when the component will receive props */
-  public static getDerivedStateFromProps(
-    nextProps: DocumentTitlePagerComponent['props'],
-    lastState: Partial<PagerState>,
-  ) {
-    const newState: Partial<PagerState> = {
-      ...lastState,
-      lastPage: nextProps.pageCount,
-    }
-    return newState
-  }
-
-  /**
-   * sets the page to the specified value
-   * @param page
-   */
-  public gotoPage(page: string | number) {
+  const gotoPage = (page: string | number) => {
     let pageInt = typeof page === 'string' ? parseInt(page, 10) : page
     if (!isNaN(pageInt)) {
       pageInt = Math.max(pageInt, 1)
-      pageInt = Math.min(pageInt, this.props.pageCount)
-      this.setState(
-        {
-          currentPage: pageInt,
-        },
-        () => {
-          this.setPage()
-        },
-      )
+      pageInt = Math.min(pageInt, props.pageCount)
+      setPage(pageInt)
     }
   }
 
-  private handleFocused() {
-    this.setState({
-      ...this.state,
-      focused: true,
-    })
-  }
-
-  private handleBlur() {
-    this.setState({
-      ...this.state,
-      focused: false,
-    })
-  }
-
-  /**
-   * renders the component
-   */
-  public render() {
-    return (
-      <ClickAwayListener onClickAway={() => this.handleBlur()}>
-        <Typography
-          onClick={() => this.handleFocused()}
-          variant="h6"
-          color="inherit"
-          style={{ cursor: 'pointer', display: 'flex', alignItems: 'center', overflow: 'hidden', margin: '0 2.5em' }}
-          title={this.props.documentName}>
-          <div style={{ whiteSpace: 'nowrap', overflow: 'hidden', textOverflow: 'ellipsis' }}>
-            {this.props.documentName}&nbsp;
+  return (
+    <ClickAwayListener onClickAway={() => setIsFocused(false)}>
+      <Typography
+        onClick={() => setIsFocused(true)}
+        variant="h6"
+        color="inherit"
+        style={{ cursor: 'pointer', display: 'flex', alignItems: 'center', overflow: 'hidden', margin: '0 2.5em' }}
+        title={props.documentName}>
+        <div style={{ whiteSpace: 'nowrap', overflow: 'hidden', textOverflow: 'ellipsis' }}>
+          {props.documentName}&nbsp;
+        </div>
+        {isFocused ? (
+          <form onSubmit={event => event.preventDefault()}>
+            <TextField
+              style={{ flexShrink: 0 }}
+              title={props.gotoPage}
+              value={props.activePages[0]}
+              onChange={ev => gotoPage(ev.currentTarget.value)}
+              type="number"
+              required={true}
+              InputLabelProps={{
+                shrink: true,
+              }}
+              inputProps={{
+                min: 1,
+                max: props.pageCount,
+                style: { textAlign: 'center' },
+              }}
+              margin="dense"
+            />
+          </form>
+        ) : (
+          <div style={{ flexShrink: 0 }}>
+            {props.activePages[0]} / {props.pageCount}
           </div>
-          {this.state.focused ? (
-            <form onSubmit={event => event.preventDefault()}>
-              <TextField
-                style={{ flexShrink: 0 }}
-                title={this.props.gotoPage}
-                defaultValue={this.state.currentPage}
-                onChange={ev => this.gotoPage(ev.currentTarget.value)}
-                type="number"
-                required={true}
-                InputLabelProps={{
-                  shrink: true,
-                }}
-                inputProps={{
-                  min: 1,
-                  max: this.state.lastPage,
-                  style: { textAlign: 'center' },
-                }}
-                margin="dense"
-              />
-            </form>
-          ) : (
-            <div style={{ flexShrink: 0 }}>
-              {this.props.activePages[0]} / {this.props.pageCount}
-            </div>
-          )}
-        </Typography>
-      </ClickAwayListener>
-    )
-  }
+        )}
+      </Typography>
+    </ClickAwayListener>
+  )
 }
 
 const connectedComponent = connect(
