@@ -8,6 +8,7 @@ export const SelectionControl: React.FunctionComponent<{ isSelected: boolean; co
   content,
 }) => (
   <div style={{ display: 'flex', alignItems: 'center', justifyContent: 'center', width: 40, height: 40 }}>
-    {isSelected ? <Checkbox checked={true} /> : <Icon item={content} />}
+    <Checkbox checked={true} style={{ display: isSelected ? undefined : 'none' }} />
+    <Icon item={content} style={{ display: isSelected ? 'none' : undefined }} />
   </div>
 )
