@@ -7,13 +7,14 @@ import { ConstantContent, ODataParams } from '@sensenet/client-core'
 import { RouteComponentProps, withRouter } from 'react-router'
 import { QueryWidget as QueryWidgetModel } from '../../services/PersonalSettings'
 import { useContentRouting, useLocalization, useRepository, useSelectionService, useStringReplace } from '../../hooks'
-import { CollectionComponent, isReferenceField } from '../ContentListPanel'
+import { CollectionComponent, isReferenceField } from '../content-list'
 import {
   CurrentAncestorsContext,
   CurrentChildrenContext,
   CurrentContentContext,
   LoadSettingsContext,
 } from '../../context'
+import { encodeQueryData } from '../search'
 
 const QueryWidget: React.FunctionComponent<QueryWidgetModel<GenericContent> & RouteComponentProps> = props => {
   const [items, setItems] = useState<GenericContent[]>([])
@@ -92,7 +93,9 @@ const QueryWidget: React.FunctionComponent<QueryWidgetModel<GenericContent> & Ro
               style={{ padding: '0', margin: '0 0 0 1em' }}
               onClick={() =>
                 props.history.push(
-                  `/${btoa(repo.configuration.repositoryUrl)}/search/${encodeURIComponent(props.settings.query)}`,
+                  `/${btoa(repo.configuration.repositoryUrl)}/search/${encodeQueryData({
+                    term: props.settings.query,
+                  })}`,
                 )
               }>
               <OpenInNewTwoTone />
@@ -138,7 +141,7 @@ const QueryWidget: React.FunctionComponent<QueryWidgetModel<GenericContent> & Ro
                     height: props.settings.countOnly || items.length < 1 ? 0 : '100%',
                   }}
                   enableBreadcrumbs={false}
-                  parentId={0}
+                  parentIdOrPath={0}
                   onParentChange={() => {
                     // props.history.push(contentRouter.getPrimaryActionUrl(p))
                   }}
