@@ -1,11 +1,15 @@
 import { Repository } from '@sensenet/client-core'
 import { usePersonalSettings } from './use-personal-settings'
 
-export const useWidgets = (repository?: Repository) => {
+export const useWidgets = (repository?: Repository, dashboardName?: string) => {
   const personalSettings = usePersonalSettings()
   if (repository) {
     const currentRepo = personalSettings.repositories.find(r => r.url === repository.configuration.repositoryUrl)
-    return currentRepo && currentRepo.dashboard ? currentRepo.dashboard : personalSettings.dashboards.repositoryDefault
+    return dashboardName && personalSettings.dashboards[dashboardName]
+      ? personalSettings.dashboards[dashboardName]
+      : currentRepo && currentRepo.dashboard
+      ? currentRepo.dashboard
+      : personalSettings.dashboards.repositoryDefault
   }
   return personalSettings.dashboards.globalDefault
 }
