@@ -61,6 +61,8 @@ const values = {
     cancelButton: 'Cancel',
     deleteSuccessNotification: `Content '{0}' has been deleted succesfully`,
     deleteMultipleSuccessNotification: `{0} content deleted succesfully`,
+    deleteSingleContentFailedNotification: `There was an error deleting content '{0}': {1}`,
+    deleteMultipleContentFailedNotification: `There was an error deleting {0} content.`,
     deleteFailedNotification: `There was an error during content deletion.`,
   },
   copyMoveContentDialog: {
@@ -88,28 +90,34 @@ const values = {
     },
   },
   drawer: {
+    titles: {
+      Content: 'Content',
+      'Content Types': 'Content Types',
+      Localization: 'Localization',
+      Search: 'Search',
+      Setup: 'Setup',
+      Trash: 'Trash',
+      'Version info': 'Version Info',
+      'Users and groups': 'Users and groups',
+    },
+    descriptions: {
+      Content: 'Explore and manage your content in the repository',
+      'Content Types': 'Manage content types',
+      Localization: 'Manage string resources',
+      Search: 'Execute custom searches, build and save queries',
+      Setup: 'Configure the sensenet system',
+      Trash: 'Manage deleted items here: restore content or purge them permanently',
+      'Version info': 'Detailed version information about the current sensenet installation',
+      'Users and groups': 'Manage users and groups, roles and identities',
+    },
     personalSettingsTitle: 'Edit personal settings',
     personalSettingsSecondaryText: 'Customize the application behavior',
-    contentTitle: 'Content',
-    contentSecondaryText: 'Explore and manage your content in the repository',
-    searchTitle: 'Search',
-    searchSecondaryText: 'Execute custom searches, build and save queries',
-    usersAndGroupsTitle: 'Users and groups',
-    usersAndGroupsSecondaryText: 'Manage users and groups, roles and identities',
-    setupTitle: 'Setup',
-    setupSecondaryText: 'Configure the sensenet system',
-    versionInfoTitle: 'Version Info',
-    versionInfoSecondaryText: 'Detailed version information about the current sensenet installation',
-    contentTypesTitle: 'Content Types',
-    contentTypesSecondaryText: 'Manage content types',
-    localizationTitle: 'Localization',
-    localizationSecondaryText: 'Manage string resources',
-    trashTitle: 'Trash',
-    trashSecondaryText: 'Manage deleted items here: restore content or purge them permanently',
+    contentRootDescription: 'The root path. Content will be displayed below this level.',
     dashboardTitle: 'Dashboard',
     dashboardSecondaryText: 'Repository overview',
     expand: 'Expand',
     collapse: 'Collapse',
+    queryTermDescription: 'The Query term',
   },
   editPropertiesDialog: {
     dialogTitle: 'Edit {0}',
@@ -117,7 +125,7 @@ const values = {
     saveFailedNotification: `There was an error during updating content '{0}'`,
   },
   login: {
-    loginTitle: 'Login',
+    loginTitle: "It's good to see you!",
     loginButtonTitle: 'Login',
     userNameLabel: 'UserName',
     userNameHelperText: "Enter the user name you've registered with",
@@ -128,6 +136,15 @@ const values = {
     loginFailed: 'Login failed.',
     greetings: 'Greetings, {0}!',
     loggingInTo: 'Logging in to {0}...',
+    youCanLogInWith: 'You can also log in with',
+    logInWithSso: 'Log in via SSO',
+    resetPassword: 'Reset password',
+    resendConfirmation: 'Resend confirmation',
+    unlockAccount: 'Unlock account',
+    newToSensenet: 'New to sensenet?',
+    help: 'Help',
+    tutorials: 'Tutorials',
+    signUp: 'Sign up',
     loginSuccessNotification: `Logged in with user '{0} to repository '{1}'`,
     loginFailedNotification: `Failed to log in with user '{0}' to repository '{1}'`,
     loginErrorNotification: `There was an error during login with user '{0}' to repository '{1}'`,
@@ -155,6 +172,9 @@ const values = {
     drawerEnable: 'Enable or disable the drawer',
     drawerType: 'Drawer type',
     drawerItems: 'Items enabled on the drawer',
+    drawerItemTitle: 'Title of the item',
+    drawerItemDescription: 'Description of the item',
+    drawerItemIcon: 'The icon of the drawer item',
     repositoryTitle: 'A list of visited repositories',
     repositoryUrl: 'The path of the repository, e.g.: https://my-sensenet-repository.org',
     repositoryLoginName: "The last user you've logged in with",
