@@ -35,24 +35,29 @@ const values: DeepPartial<typeof import('./default').default> = {
     cancelButton: 'Mégsem',
   },
   drawer: {
-    contentTitle: 'Tartalom',
-    contentSecondaryText: 'Tartalom böngészése',
-    searchTitle: 'Keresés',
-    searchSecondaryText: 'Testreszabott keresések futtatása és mentése későbbi használatra',
-    setupTitle: 'Beállítás',
-    setupSecondaryText: 'A rendszer beállításai',
-    usersAndGroupsTitle: 'Felhasználók és csoportok',
-    usersAndGroupsSecondaryText: 'Felhasználó és csoport kezelése, szerkesztése',
-    versionInfoTitle: 'Verzió névjegye',
-    versionInfoSecondaryText: 'Információk a telepített csomagokról és a verzióikról',
+    titles: {
+      Content: 'Tartalom',
+      Search: 'Keresés',
+      Setup: 'Beállítás',
+      'Users and groups': 'Felhasználók és csoportok',
+      'Version info': 'Verzió névjegye',
+      'Content Types': 'Tartalom típusok',
+      Localization: 'Nyelvi fájlok',
+      Trash: 'Kuka',
+    },
+    descriptions: {
+      Content: 'Tartalom böngészése',
+      Search: 'Testreszabott keresések futtatása és mentése későbbi használatra',
+      Setup: 'A rendszer beállításai',
+      'Users and groups': 'Felhasználó és csoport kezelése, szerkesztése',
+      'Version info': 'Információk a telepített csomagokról és a verzióikról',
+      'Content Types': 'Tartalom típusok kezelése',
+      Localization: 'Nyelvi fájlok kezelése',
+      Trash: 'Törölt elemek kezelése',
+    },
+
     personalSettingsTitle: 'Személyes beállítások',
     personalSettingsSecondaryText: 'Az alkalmazás testreszabása',
-    contentTypesTitle: 'Tartalom típusok',
-    contentTypesSecondaryText: 'Tartalom típusok kezelése',
-    localizationTitle: 'Nyelvi fájlok',
-    localizationSecondaryText: 'Nyelvi fájlok kezelése',
-    trashTitle: 'Kuka',
-    trashSecondaryText: 'Törölt elemek kezelése',
     dashboardTitle: 'Irányítópult',
     dashboardSecondaryText:
       'Olyan egyoldalas vizuális felület, amelynek segítségével a felhasználó első ránézésre monitorozhatja legfontosabb céljainak vagy elvárásainak megvalósulását',
