@@ -15,7 +15,7 @@ describe('ODataUrlBuilder', () => {
     it('should return an empty string, if the argument is undefined', () => {
       const urlParamString = ODataUrlBuilder.buildUrlParamString<GenericContent>(repo.configuration)
       expect(urlParamString).toBe(
-        '$select=Id,Path,Name,Type,DisplayName,Description,Icon&metadata=no&$inlinecount=allpages&$top=10000',
+        '$select=Id%2CPath%2CName%2CType%2CDisplayName%2CDescription%2CIcon&metadata=no&$inlinecount=allpages&$top=10000',
       )
     })
     it("should return a string with only select Id and Type if there's no selected field", () => {
@@ -23,14 +23,14 @@ describe('ODataUrlBuilder', () => {
         { defaultSelect: ['Id', 'Type'] } as any,
         { metadata: 'no' },
       )
-      expect(urlParamString).toBe('metadata=no&$select=Id,Type')
+      expect(urlParamString).toBe('metadata=no&$select=Id%2CType')
     })
     it('should return a string with the given field and Id and Type as selected', () => {
       const urlParamString = ODataUrlBuilder.buildUrlParamString<GenericContent>(
         { requiredSelect: ['Id', 'Type'], defaultMetadata: 'no' } as any,
         { select: 'DisplayName' },
       )
-      expect(urlParamString).toBe('$select=Id,Type,DisplayName&metadata=no')
+      expect(urlParamString).toBe('$select=Id%2CType%2CDisplayName&metadata=no')
     })
     it('should return a string without select if requiredSelect is all', () => {
       const urlParamString = ODataUrlBuilder.buildUrlParamString<GenericContent>(
@@ -44,14 +44,14 @@ describe('ODataUrlBuilder', () => {
         { requiredSelect: ['Id', 'Type'], defaultMetadata: 'no' } as any,
         { select: ['DisplayName', 'Path'] },
       )
-      expect(urlParamString).toBe('$select=Id,Type,DisplayName,Path&metadata=no')
+      expect(urlParamString).toBe('$select=Id%2CType%2CDisplayName%2CPath&metadata=no')
     })
     it('should return a string with the given parameters', () => {
       const urlParamString = ODataUrlBuilder.buildUrlParamString<GenericContent>(
         { requiredSelect: ['Id', 'Type'], defaultMetadata: 'no' } as any,
         { select: ['DisplayName', 'Path'], orderby: 'DisplayName' },
       )
-      expect(urlParamString).toBe('$select=Id,Type,DisplayName,Path&$orderby=DisplayName&metadata=no')
+      expect(urlParamString).toBe('$select=Id%2CType%2CDisplayName%2CPath&$orderby=DisplayName&metadata=no')
     })
     it('should return a string with the given parameters', () => {
       const urlParamString = ODataUrlBuilder.buildUrlParamString<GenericContent>(
@@ -62,7 +62,9 @@ describe('ODataUrlBuilder', () => {
           query: new Query(q => q.typeIs<Task>(Task)).toString(),
         },
       )
-      expect(urlParamString).toBe('$select=Id,Type,DisplayName,Path&$orderby=DisplayName&query=TypeIs:Task&metadata=no')
+      expect(urlParamString).toBe(
+        '$select=Id%2CType%2CDisplayName%2CPath&$orderby=DisplayName&query=TypeIs%3ATask&metadata=no',
+      )
     })
     it('should return a string without select param', () => {
       const urlParamString = ODataUrlBuilder.buildUrlParamString<GenericContent>({ defaultMetadata: 'no' } as any, {
@@ -82,28 +84,28 @@ describe('ODataUrlBuilder', () => {
       const urlParamString = ODataUrlBuilder.buildUrlParamString<GenericContent>({ defaultMetadata: 'no' } as any, {
         orderby: ['Name', 'DisplayName'],
       })
-      expect(urlParamString).toBe('$orderby=Name,DisplayName&metadata=no')
+      expect(urlParamString).toBe('$orderby=Name%2CDisplayName&metadata=no')
     })
 
     it('should parse an orderby field expression with order', () => {
       const urlParamString = ODataUrlBuilder.buildUrlParamString<GenericContent>({ defaultMetadata: 'no' } as any, {
         orderby: [['Name', 'asc']],
       })
-      expect(urlParamString).toBe('$orderby=Name asc&metadata=no')
+      expect(urlParamString).toBe('$orderby=Name%20asc&metadata=no')
     })
 
     it('should parse an orderby array with ordered fields list expression', () => {
       const urlParamString = ODataUrlBuilder.buildUrlParamString<GenericContent>({ defaultMetadata: 'no' } as any, {
         orderby: [['Name', 'asc'], ['DisplayName', 'desc']],
       })
-      expect(urlParamString).toBe('$orderby=Name asc,DisplayName desc&metadata=no')
+      expect(urlParamString).toBe('$orderby=Name%20asc%2CDisplayName%20desc&metadata=no')
     })
 
     it('should parse an orderby array with ordered fields list expression and field names', () => {
       const urlParamString = ODataUrlBuilder.buildUrlParamString<GenericContent>({ defaultMetadata: 'no' } as any, {
         orderby: [['Name', 'asc'], 'DisplayName'],
       })
-      expect(urlParamString).toBe('$orderby=Name asc,DisplayName&metadata=no')
+      expect(urlParamString).toBe('$orderby=Name%20asc%2CDisplayName&metadata=no')
     })
 
     it('should return a string without any param', () => {
@@ -111,7 +113,7 @@ describe('ODataUrlBuilder', () => {
         requiredSelect: ['Id', 'Type'],
         defaultMetadata: 'no',
       } as any)
-      expect(urlParamString).toBe('$select=Id,Type&metadata=no')
+      expect(urlParamString).toBe('$select=Id%2CType&metadata=no')
     })
   })
 })
