@@ -1,7 +1,8 @@
 import { FieldSetting, GenericContent } from '@sensenet/default-content-types'
 import { ContentListProps } from '../ContentList'
 
-export interface CellProps<T extends GenericContent, K extends keyof T> extends ContentListProps<T> {
+export interface CellProps<T extends GenericContent = GenericContent, K extends keyof T = keyof GenericContent>
+  extends ContentListProps<T> {
   content: T
   field: K
   isSelected: boolean
