@@ -7,14 +7,14 @@ import TableRow from '@material-ui/core/TableRow'
 import TableSortLabel from '@material-ui/core/TableSortLabel'
 import Tooltip from '@material-ui/core/Tooltip'
 
-import { ActionModel, GenericContent, Schema } from '@sensenet/default-content-types'
-import React from 'react'
+import { ActionModel, FieldSetting, GenericContent, Schema } from '@sensenet/default-content-types'
+import React, { useCallback, useMemo } from 'react'
 import { ActionsCell, CellProps, DateCell, DefaultCell, DisplayNameCell, ReferenceCell } from './CellTemplates'
 
 /**
  * Interface for ContentList properties
  */
-export interface ContentListProps<T extends GenericContent> {
+export interface ContentListProps<T extends GenericContent = GenericContent> {
   /**
    * Array of content items
    */
@@ -112,196 +112,175 @@ export interface ContentListProps<T extends GenericContent> {
   hideHeader?: boolean
 }
 
-export interface ContentListState {
-  itemCount: number
-  selectedCount: number
-  isAllSelected: boolean
-  hasSelected: boolean
-}
+export const ContentList: React.FC<ContentListProps<GenericContent>> = props => {
+  const handleSelectAllClick = useCallback(() => {
+    props.onRequestSelectionChange &&
+      (props.selected && props.selected.length === props.items.length
+        ? props.onRequestSelectionChange([])
+        : props.onRequestSelectionChange(props.items))
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [props.onRequestSelectionChange, props.selected, props.items])
 
-export class ContentList<T extends GenericContent> extends React.Component<ContentListProps<T>, ContentListState> {
-  public state = ContentList.getDerivedStateFromProps(this.props as any)
-
-  public static getDerivedStateFromProps(nextProps: ContentListProps<GenericContent>) {
-    const selected = nextProps.selected ? nextProps.selected : []
-    const selectedCount = selected.length
-    const itemCount = nextProps.items.length
-    return {
-      selectedCount,
-      itemCount,
-      hasSelected: selectedCount > 0,
-      isAllSelected: itemCount === selectedCount,
-    }
-  }
-
-  public handleSelectAllClick() {
-    this.props.onRequestSelectionChange &&
-      (this.state.isAllSelected
-        ? this.props.onRequestSelectionChange([])
-        : this.props.onRequestSelectionChange(this.props.items))
-  }
-
-  public handleContentSelection(content: T) {
-    const selected = this.props.selected !== undefined && this.props.selected.length > 0 ? this.props.selected : []
-    if (this.props.onRequestSelectionChange) {
-      if (selected.find(c => c.Id === content.Id)) {
-        this.props.onRequestSelectionChange(selected.filter(s => s.Id !== content.Id))
-      } else {
-        this.props.onRequestSelectionChange([...selected, content])
-      }
-    }
-  }
-
-  private defaultFieldComponents: React.StatelessComponent<CellProps<T, keyof T>> = (props: CellProps<T, keyof T>) => {
-    switch (props.field) {
-      case 'DisplayName':
-        return <DisplayNameCell content={props.content} isSelected={props.isSelected} icons={this.props.icons} />
-      case 'Actions':
-        if (props.content.Actions && props.content.Actions instanceof Array) {
-          return (
-            <ActionsCell
-              actions={props.content.Actions as ActionModel[]}
-              content={props.content}
-              openActionMenu={ev =>
-                this.props.onRequestActionsMenu && this.props.onRequestActionsMenu(ev, props.content)
-              }
-            />
-          )
+  const handleContentSelection = useCallback(
+    (content: GenericContent) => {
+      const selected = props.selected !== undefined && props.selected.length > 0 ? props.selected : []
+      if (props.onRequestSelectionChange) {
+        if (selected.find(c => c.Id === content.Id)) {
+          props.onRequestSelectionChange(selected.filter(s => s.Id !== content.Id))
+        } else {
+          props.onRequestSelectionChange([...selected, content])
         }
-        break
-      case 'ModificationDate':
-        return <DateCell date={props.content.ModificationDate as string} />
-      default:
-        break
-    }
-    const field: any = props.content[props.field]
-    if (field && field.Id && field.Path && field.DisplayName) {
-      return <ReferenceCell content={field} fieldName={'DisplayName'} />
-    }
-    return null
-  }
+      }
+    },
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [props.onRequestSelectionChange, props.selected, props.items],
+  )
 
-  constructor(props: ContentListProps<T>) {
-    super(props)
-    this.handleSelectAllClick = this.handleSelectAllClick.bind(this)
-    this.handleContentSelection = this.handleContentSelection.bind(this)
-  }
-  public render() {
-    const selected = this.props.selected ? this.props.selected : []
-    const orderDirection = this.props.orderDirection ? this.props.orderDirection : 'asc'
-    const orderBy = this.props.orderBy ? this.props.orderBy : 'DisplayName'
-    return (
-      <Table>
-        {this.props.hideHeader ? null : (
-          <TableHead>
-            <TableRow>
-              {this.props.displayRowCheckbox !== false ? (
-                <TableCell padding="checkbox" key="selectAll" style={{ width: '30px', paddingRight: 0 }}>
-                  <Checkbox
-                    className="select-all"
-                    indeterminate={this.state.hasSelected && !this.state.isAllSelected}
-                    checked={this.state.isAllSelected}
-                    onChange={this.handleSelectAllClick}
-                  />
+  const defaultFieldComponents: React.FC<CellProps<GenericContent, keyof GenericContent>> = useCallback(
+    fieldProps => {
+      switch (fieldProps.field) {
+        case 'DisplayName':
+          return <DisplayNameCell content={fieldProps.content} isSelected={fieldProps.isSelected} icons={props.icons} />
+        case 'Actions':
+          if (fieldProps.content.Actions && fieldProps.content.Actions instanceof Array) {
+            return (
+              <ActionsCell
+                actions={fieldProps.content.Actions as ActionModel[]}
+                content={fieldProps.content}
+                openActionMenu={ev =>
+                  fieldProps.onRequestActionsMenu && fieldProps.onRequestActionsMenu(ev, fieldProps.content)
+                }
+              />
+            )
+          }
+          break
+        case 'ModificationDate':
+          return <DateCell date={fieldProps.content.ModificationDate as string} />
+        default:
+          break
+      }
+      const field: any = fieldProps.content[fieldProps.field]
+      if (field && field.Id && field.Path && field.DisplayName) {
+        return <ReferenceCell content={field} fieldName={'DisplayName'} />
+      }
+      return <DefaultCell {...fieldProps} />
+    },
+    [props.icons],
+  )
+
+  const fieldSchemas = useMemo<{ [key: string]: FieldSetting }>(
+    () =>
+      props.schema.FieldSettings.reduce((v, field) => {
+        ;(v as any)[field.Name] = props.schema.FieldSettings.find(s => s.Name === field.Name)
+        return v
+      }, {}),
+    [props.schema.FieldSettings],
+  )
+
+  const getSchemaForField = useCallback((fieldName: string) => fieldSchemas[fieldName] as FieldSetting, [fieldSchemas])
+
+  return (
+    <Table>
+      {props.hideHeader ? null : (
+        <TableHead>
+          <TableRow>
+            {props.displayRowCheckbox !== false ? (
+              <TableCell padding="checkbox" key="selectAll" style={{ width: '30px', paddingRight: 0 }}>
+                <Checkbox
+                  className="select-all"
+                  indeterminate={
+                    props.selected && props.selected.length > 0 && props.selected.length !== props.items.length
+                      ? true
+                      : false
+                  }
+                  checked={props.selected && props.selected.length === props.items.length ? true : false}
+                  onChange={handleSelectAllClick}
+                />
+              </TableCell>
+            ) : null}
+            {props.fieldsToDisplay
+              ? props.fieldsToDisplay.map(field => {
+                  const fieldSetting = getSchemaForField(field)
+                  const isNumeric =
+                    fieldSetting &&
+                    (fieldSetting.Type === 'IntegerFieldSetting' || fieldSetting.Type === 'NumberFieldSetting')
+                  const description = (fieldSetting && fieldSetting.Description) || field
+                  const displayName = (fieldSetting && fieldSetting.DisplayName) || field
+                  return (
+                    <TableCell
+                      key={field as string}
+                      align={isNumeric ? 'right' : 'inherit'}
+                      className={field as string}>
+                      <Tooltip title={description}>
+                        <TableSortLabel
+                          active={props.orderBy === field}
+                          direction={props.orderDirection}
+                          onClick={() =>
+                            props.onRequestOrderChange &&
+                            props.onRequestOrderChange(field, props.orderDirection === 'asc' ? 'desc' : 'asc')
+                          }>
+                          {displayName}
+                        </TableSortLabel>
+                      </Tooltip>
+                    </TableCell>
+                  )
+                })
+              : null}
+          </TableRow>
+        </TableHead>
+      )}
+      <TableBody>
+        {props.items.map(item => {
+          const isSelected = props.selected && props.selected.find(s => s.Id === item.Id) ? true : false
+          const isActive = props.active && props.active.Id === item.Id ? true : false
+          return (
+            <TableRow
+              key={item.Id}
+              hover={true}
+              className={`${isActive ? 'active' : ''} ${isSelected ? 'selected' : ''} ${item.Type &&
+                `type-${item.Type.toLowerCase()}`}`}
+              selected={isActive}
+              onClick={e => {
+                props.onRequestActiveItemChange && props.onRequestActiveItemChange(item)
+                props.onItemClick && props.onItemClick(e, item)
+              }}
+              onDoubleClick={e => {
+                props.onItemDoubleClick && props.onItemDoubleClick(e, item)
+              }}
+              onTouchEnd={e => props.onItemTap && props.onItemTap(e, item)}
+              onContextMenu={e => props.onItemContextMenu && props.onItemContextMenu(e, item)}>
+              {props.displayRowCheckbox !== false ? (
+                <TableCell padding="checkbox" key="select">
+                  {props.getSelectionControl ? (
+                    props.getSelectionControl(isSelected, item)
+                  ) : (
+                    <Checkbox
+                      checked={isSelected ? true : false}
+                      onChange={() => handleContentSelection(item)}
+                      {...props.checkboxProps}
+                    />
+                  )}
                 </TableCell>
               ) : null}
-              {this.props.fieldsToDisplay
-                ? this.props.fieldsToDisplay.map(field => {
-                    const fieldSetting = this.props.schema.FieldSettings.find(s => s.Name === field)
-                    const isNumeric =
-                      fieldSetting &&
-                      (fieldSetting.Type === 'IntegerFieldSetting' || fieldSetting.Type === 'NumberFieldSetting')
-                    const description = (fieldSetting && fieldSetting.Description) || field
-                    const displayName = (fieldSetting && fieldSetting.DisplayName) || field
-                    return (
-                      <TableCell
-                        key={field as string}
-                        align={isNumeric ? 'right' : 'inherit'}
-                        className={field as string}>
-                        <Tooltip title={description}>
-                          <TableSortLabel
-                            active={orderBy === field}
-                            direction={orderDirection}
-                            onClick={() =>
-                              this.props.onRequestOrderChange &&
-                              this.props.onRequestOrderChange(field, orderDirection === 'asc' ? 'desc' : 'asc')
-                            }>
-                            {displayName}
-                          </TableSortLabel>
-                        </Tooltip>
-                      </TableCell>
-                    )
+              {props.fieldsToDisplay
+                ? props.fieldsToDisplay.map(field => {
+                    const fieldSetting = getSchemaForField(field)
+                    const cellProps: CellProps = {
+                      ...(props as ContentListProps),
+                      field,
+                      content: item,
+                      fieldSetting,
+                      isSelected,
+                    }
+
+                    const FieldComponent = props.fieldComponent || defaultFieldComponents
+                    return <FieldComponent key={cellProps.field} {...cellProps} />
                   })
                 : null}
             </TableRow>
-          </TableHead>
-        )}
-        <TableBody>
-          {this.props.items.map(item => {
-            const isSelected = selected.find(s => s.Id === item.Id) ? true : false
-            const isActive = this.props.active && this.props.active.Id === item.Id ? true : false
-            return (
-              <TableRow
-                key={item.Id}
-                hover={true}
-                className={`${isActive ? 'active' : ''} ${isSelected ? 'selected' : ''} ${item.Type &&
-                  `type-${item.Type.toLowerCase()}`}`}
-                selected={isActive}
-                onClick={e => {
-                  this.props.onRequestActiveItemChange && this.props.onRequestActiveItemChange(item)
-                  this.props.onItemClick && this.props.onItemClick(e, item)
-                }}
-                onDoubleClick={e => {
-                  this.props.onItemDoubleClick && this.props.onItemDoubleClick(e, item)
-                }}
-                onTouchEnd={e => this.props.onItemTap && this.props.onItemTap(e, item)}
-                onContextMenu={e => this.props.onItemContextMenu && this.props.onItemContextMenu(e, item)}>
-                {this.props.displayRowCheckbox !== false ? (
-                  <TableCell padding="checkbox" key="select">
-                    {this.props.getSelectionControl ? (
-                      this.props.getSelectionControl(selected.find(i => i.Id === item.Id) ? true : false, item)
-                    ) : (
-                      <Checkbox
-                        checked={selected.find(i => i.Id === item.Id) ? true : false}
-                        onChange={() => this.handleContentSelection(item)}
-                        {...this.props.checkboxProps}
-                      />
-                    )}
-                  </TableCell>
-                ) : null}
-                {this.props.fieldsToDisplay
-                  ? this.props.fieldsToDisplay.map(field => {
-                      const fieldSetting = this.props.schema.FieldSettings.find(s => s.Name === field)
-                      const cellProps: CellProps<T, keyof T> = {
-                        ...(this.props as ContentListProps<T>),
-                        field,
-                        content: item,
-                        fieldSetting,
-                        isSelected,
-                      }
-
-                      const fieldComponent = this.props.fieldComponent && this.props.fieldComponent(cellProps)
-                      const defaultComponent = this.defaultFieldComponents(cellProps)
-
-                      const el = {
-                        ...React.createElement(
-                          fieldComponent
-                            ? (this.props.fieldComponent as any)
-                            : defaultComponent
-                            ? this.defaultFieldComponents
-                            : DefaultCell,
-                          cellProps,
-                        ),
-                        key: field as string,
-                      }
-                      return el
-                    })
-                  : null}
-              </TableRow>
-            )
-          })}
-        </TableBody>
-      </Table>
-    )
-  }
+          )
+        })}
+      </TableBody>
+    </Table>
+  )
 }
